+++ R/utils-upload.R
@@ -10,8 +10,11 @@
 #' document is completely self-contained, i.e. it does not need external image
 #' files any more, and it is ready to be published online.
 #' @param file Path to the image file to be uploaded.
-#' @param key Client ID for Imgur. By default, this uses a client ID registered
-#'   by Yihui Xie.
+#' @param key Client ID for Imgur. It can be set via either the global option
+#'   \code{knitr.imgur.key} (you may set it in \file{~/.Rprofile}) or the
+#'   environment variable \code{R_KNITR_IMGUR_KEY} (you may set it in
+#'   \file{~/.Renviron}). If neither is set, this uses a client ID registered by
+#'   Yihui Xie.
 #' @return A character string of the link to the image; this string carries an
 #'   attribute named \code{XML} which is a list converted from the response XML
 #'   file; see Imgur API in the references.
@@ -31,9 +34,13 @@
 #' if (interactive()) browseURL(res)
 #'
 #' # to use your own key
-#' opts_knit$set(upload.fun = function(file) imgur_upload(file, key = 'your imgur key'))
+#' options(knitr.imgur.key = 'your imgur key')
 #' }
-imgur_upload = function(file, key = '9f3460e67f308f6') {
+imgur_upload = function(
+  file, key = getOption(
+    'knitr.imgur.key', Sys.getenv('R_KNITR_IMGUR_KEY', '9f3460e67f308f6')
+  )
+) {
   if (!is.character(key)) stop('The Imgur API Key must be a character string!')
   resp = httr::POST(
     "https://api.imgur.com/3/image.xml",